insurance price & percent added

+++ cmd/api/internal/storage/kasko.go
@@ -4,10 +4,10 @@ import "gorm.io/gorm"
 
 type Kasko struct {
 	gorm.Model
-	Title       string `gorm:"size:255;" json:"title"`
-	Price       int    `json:"price"`
-	Percent     int    `json:"percent"`
-	InsuranceID uint   `json:"insuranceID"`
+	Title       string  `gorm:"size:255;" json:"title"`
+	Price       int     `json:"price"`
+	Percent     float64 `json:"percent"`
+	InsuranceID uint    `json:"insuranceID"`
 }
 
 func (k *Kasko) Save(db *gorm.DB) (*Kasko, error) {
